Create Alice and Strings.py
Two strings  and  comprising of lower case English letters are compatible if they are equal or can be made equal by following this step any number of times:

Select a prefix from the string  (possibly empty), and increase the alphabetical value of all the characters in the prefix by the same valid amount. For example if the string is  and we select the prefix  then we can convert it to  by increasing the alphabetical value by 1. But if we select the prefix  then we cannot increase the alphabetical value.